Encrypt password
Created auth controller and auth services,
Installed bcrypt,
